See the extended description.
When in, or going to be in, PALmode, we need to use the PALshadow registers for certain integer registers (4-7, 20-23).  All other registers, including all the Floating-point ones, do not get changed.
The calculation of the PALcode entry point was not correct.  It was not taking bits 5-0 out of the function code and adding it to the calculated PC.
Support was added to CPUs prior to the 21264, this includes the 21164 and 21064 CPUs prior, and the 21364 and 21464 after, for calculating the PALcode entry point.